backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
volcengine==1.0.0
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.12.0
aiofiles==23.2.1
requests==2.31.0
qwen-agent>=0.0.8
mcp>=0.1.0

